requirements.txt
snowflake-connector-python
boto3~=1.34.142
Flask~=3.0.3
pandas~=2.2.2
werkzeug~=3.0.3
flask_mail
pysnowflake~=0.1.3
botocore~=1.34.142